tidy-up and version bumps
