+++ .editorconfig
@@ -1,6 +1,7 @@
 root = true
 
-[*]
+[*.{kt,kts}]
 indent_size = 2
 continuation_indent_size = 4
 ij_kotlin_name_count_to_use_star_import = unset
+disabled_rules = no-wildcard-imports
