+++ buildSrc/settings.gradle.kts
@@ -1,6 +1,6 @@
 pluginManagement {
   plugins {
-    id("de.fayard.refreshVersions") version "0.23.0"
+    id("de.fayard.refreshVersions") version "0.30.1"
   }
 }
