+++ buildSrc/src/main/kotlin/plugin.common.gradle.kts
@@ -1,3 +1,4 @@
+import de.fayard.refreshVersions.core.versionFor
 import org.jetbrains.kotlin.gradle.tasks.AbstractKotlinCompile
 import org.jetbrains.kotlin.gradle.tasks.KotlinTest
 
@@ -20,8 +21,19 @@ idea {
 }
 
 spotless {
-  kotlin { ktfmt() }
-  kotlinGradle { ktfmt() }
+  val ktlintSettings = mapOf(
+    "indent_size" to "2",
+    "continuation_indent_size" to "4",
+    "disabled_rules" to "no-wildcard-imports"
+  )
+  kotlin {
+    target("src/**/*.kt")
+    ktlint(versionFor("version.ktlint")).userData(ktlintSettings)
+  }
+  kotlinGradle {
+    target("**/*.kts")
+    ktlint(versionFor("version.ktlint")).userData(ktlintSettings)
+  }
 }
 
 tasks {
