+++ buildSrc/src/main/kotlin/plugin.git-hooks.gradle.kts
@@ -1,3 +1,7 @@
-plugins { id("com.github.jakemarsden.git-hooks") }
+plugins {
+  id("com.github.jakemarsden.git-hooks")
+}
 
-gitHooks { setHooks(mapOf("pre-commit" to "spotlessApply", "pre-push" to "spotlessCheck")) }
+gitHooks {
+  setHooks(mapOf("pre-commit" to "spotlessApply", "pre-push" to "spotlessCheck"))
+}
