+++ docs/CONTRIBUTING.md
@@ -1,0 +1,2 @@
+# Contributing to KMDC
+TODO
