+++ lib/kmdc-card/src/jsMain/kotlin/MDCCardActions.kt
@@ -1,11 +1,11 @@
 package dev.petuska.kmdc.card
 
 import androidx.compose.runtime.Composable
+import dev.petuska.kmdc.button.MDCButton
+import dev.petuska.kmdc.button.MDCButtonScope
 import dev.petuska.kmdc.core.Builder
 import dev.petuska.kmdc.core.ComposableBuilder
 import dev.petuska.kmdc.core.MDCDsl
-import dev.petuska.kmdc.button.MDCButton
-import dev.petuska.kmdc.button.MDCButtonScope
 import dev.petuska.kmdc.icon.button.MDCIconButton
 import dev.petuska.kmdc.icon.button.MDCIconButtonScope
 import dev.petuska.kmdc.icon.button.MDCIconLink
