+++ sandbox/build.gradle.kts
@@ -1,8 +1,8 @@
-import org.jetbrains.kotlin.gradle.targets.js.webpack.KotlinWebpackConfig
+
 
 plugins {
-  kotlin("multiplatform") version "1.5.31"
-  id("org.jetbrains.compose") version "1.0.0-rc1"
+  kotlin("multiplatform") version "1.6.10"
+  id("org.jetbrains.compose") version "1.0.1-rc2"
 }
 
 repositories {
@@ -39,12 +39,5 @@ kotlin {
         implementation("dev.petuska:kmdc")
       }
     }
-  }
-}
-
-// Workaround for https://kotlinlang.slack.com/archives/C0B8L3U69/p1633590092096600
-rootProject.plugins.withType<org.jetbrains.kotlin.gradle.targets.js.yarn.YarnPlugin> {
-  rootProject.the<org.jetbrains.kotlin.gradle.targets.js.yarn.YarnRootExtension>().apply {
-    resolution("@webpack-cli/serve", "1.5.2")
   }
 }
