+++ scripts/removeGhPackageVersion.main.kts
@@ -87,7 +87,6 @@ runBlocking {
       }
       toList()
     }
-
   }
   versions.forEach { (pkg, ver) ->
     if (ver.size > 1) {
