+++ versions.properties
@@ -1,21 +1,33 @@
 #### Dependencies and Plugin versions with their available updates.
-#### Generated by `./gradlew refreshVersions` version 0.23.0
+#### Generated by `./gradlew refreshVersions` version 0.30.1
 ####
 #### Don't manually edit or split the comments that start with four hashtags (####),
 #### they will be overwritten by refreshVersions.
 ####
 #### suppress inspection "SpellCheckingInspection" for whole file
 #### suppress inspection "UnusedProperty" for whole file
+
 #====================================================== Plugins ========================================================
-version.com.diffplug.spotless..spotless-plugin-gradle=6.0.0
+version.com.diffplug.spotless..spotless-plugin-gradle=6.0.5
+
 version.com.github.jakemarsden..git-hooks-gradle-plugin=0.0.2
+
 version.io.github.gradle-nexus..publish-plugin=1.1.0
+
 version.org.jetbrains.dokka..dokka-gradle-plugin=1.6.0
+
+version.ktlint=0.43.2
+
 #===================================================== Libraries =======================================================
-version.klip=0.2.2
-version.kotlin=1.5.31
-version.kotlinx.coroutines=1.5.2
-version.org.jetbrains.compose..compose-gradle-plugin=1.0.0
+ version.klip=0.3.0
+
+version.kotlin=1.6.10
+
+version.kotlinx.coroutines=1.6.0
+
+version.org.jetbrains.compose..compose-gradle-plugin=1.0.1-rc2
+
 #======================================================== NPM ==========================================================
 version.npm.material-components-web=^13.0.0
+
 version.npm.material-icons=^1.10.4
